backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
joblib==1.5.0
numpy==2.0.2
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
